requirements.txt
# Dependencies required for the Elysium project
# Telegram Bot Dependencies
python-telegram-bot==13.7
urllib3==1.26.15
httpx==0.23.0

# API and Network Dependencies
requests>=2.28.2
aiohttp>=3.8.4

# Cryptography and Security
cryptography>=39.0.2
pycryptodome>=3.17.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
jsonschema>=4.17.3

# For development
pytest>=7.3.1
black>=23.3.0
isort>=5.12.0
mypy>=1.3.0
pylint>=2.17.4